[android] MainActivity/NavigationMenu refactors.
 - NavigationMenu moved to com.frostwire.android.gui/activities.internal package.
 - Removed unnecessary circular references to NavigationMenu on MenuDrawerToggle's constructor
 - MenuDrawerToggle no longer static so it can directly access the NavigationMenu instance that creates it
 - More fields in NavigationMenu now final
